+++ lib/domain/firebase_options.dart
