+++ lib/domain/firestore_data_provider.dart
@@ -2,7 +2,7 @@ import 'dart:typed_data';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_storage/firebase_storage.dart';
-import 'package:genius/domain_layer/business/user.dart';
+import 'package:genius/domain/business/user.dart';
 
 class FirebaseDataProvider {
   Future<User?> readUser(String userId) async {
